# Replace ownEnumerableKeys with for-in loop 