Inspired by #20.

I inlined `ownEnumerableKeys` - because it was main bottleneck of performance (heavy `filter` and `keys` construction).

It can be better, than in lodash if we disable Symbols checking, but this will break specification (in NodeJS 0.10 it will be faster thou).

IoJS 3.0.0:

```
Before changes:

                      object-assign
         459,622 op/s » small
         236,813 op/s » default options
         136,718 op/s » big

After changes:

                      object-assign
       1,280,370 op/s » small
         651,416 op/s » default options
         238,518 op/s » big
```

In Node 0.10.38:

```
Before changes:

                      object-assign
         740,634 op/s » small
         336,369 op/s » default options
          76,365 op/s » big

After changes:

                      object-assign
       4,002,658 op/s » small
       1,418,239 op/s » default options
         170,629 op/s » big
```

Any comments appreciated.