@floatdrop 2015-08-15
In Node 0.10.38:

```
Before changes:

                      object-assign
         740,634 op/s » small
         336,369 op/s » default options
          76,365 op/s » big

                      lodash
       2,180,606 op/s » small
         850,214 op/s » default options
         234,963 op/s » big

After changes:

                      object-assign
       4,002,658 op/s » small
       1,418,239 op/s » default options
         170,629 op/s » big

                      lodash
       2,129,345 op/s » small
         848,346 op/s » default options
         239,885 op/s » big
```