@floatdrop 2015-08-18
@sindresorhus bump.