@mridgway 2015-08-18
Nice, this fixes #22 as well. :+1: