@sindresorhus 2015-08-20
Sorry about the delay @floatdrop. Looks great!

Some more resources just to make sure we've done it right this time:

- Spec: http://www.ecma-international.org/ecma-262/6.0/#sec-object.assign
- V8 implementation: https://chromium.googlesource.com/v8/v8.git/+/fda20efb2f75a11b711501c3d8bb19cef4873200%5E%21/#F3