@sindresorhus 2015-08-20
Note to self: This needs to be a major release as it alters the behaviour.